Move test logic to campaign repo tests
